+++ tool/ood-gen/lib/utils.ml
@@ -40,7 +40,13 @@ let read_from_dir dir =
          else
            None)
 
-let map_files f dir = read_from_dir dir |> List.map (fun (_, x) -> f x)
+let map_files f dir =
+  read_from_dir dir
+  |> List.map (fun (file, x) ->
+         try f x with
+         | exn ->
+           prerr_endline ("Error in " ^ file);
+           raise exn)
 
 let map_files_with_names f dir = read_from_dir dir |> List.map f